Create separate files for db and routes
