+++ db.py
@@ -1,0 +1,6 @@
+from os import getenv
+from flask_sqlalchemy import SQLAlchemy
+from app import app
+
+app.config["SQLALCHEMY_DATABASE_URI"] = getenv("DATABASE_URL")
+db = SQLAlchemy(app)
